Add props to ContainerUtils.ts
